Began implementing physics with Characters. Began creating a Player object for the Game. Edited the XNARenderer to layer based on floats (the greater the value, the closer to the front the object will be).
git-svn-id: https://subversion.assembla.com/svn/oak@37 083f587b-f4ff-4b6b-aeb3-7530dcb08118

+++ trunk/OakEngine/Engine/Graphics/IRenderer.cs
@@ -16,7 +16,7 @@ namespace Oak.Engine.Graphics
         public float rotation;
         public Vector2 origin;
         public SpriteEffects effect;
-        public int layerDepth;
+        public float layerDepth;
     }
 
     interface IRenderer
